+++ imod/__init__.py
@@ -14,4 +14,4 @@ from imod import (
 )
 from imod.formats import gen, idf, ipf, prj, rasterio
 
-__version__ = "1.0.0.post1"
+__version__ = "1.0.1dev1"