+++ imod/tests/test_mf6/test_mf6_evt.py
@@ -9,9 +9,7 @@ import imod
 from imod.schemata import ValidationError
 
 
-def test_render(
-    rate_fc, elevation_fc, concentration_fc, proportion_rate_fc, proportion_depth_fc
-):
+def test_render_simple(rate_fc, elevation_fc, concentration_fc):
     # Arrange
     directory = pathlib.Path("mymodel")
     globaltimes = np.array(
@@ -27,8 +25,6 @@ def test_render(
         surface=elevation_fc,
         rate=rate_fc,
         depth=elevation_fc,
-        proportion_rate=proportion_rate_fc,
-        proportion_depth=proportion_depth_fc,
         concentration=concentration_fc,
         concentration_boundary_type="AUX",
     )
@@ -61,9 +57,7 @@ def test_render(
     assert actual == expected
 
 
-def test_get_options__no_segments(
-    rate_fc, elevation_fc, proportion_rate_fc, proportion_depth_fc
-):
+def test_get_options__no_segments(rate_fc, elevation_fc):
     """Test with no segments specified, this means there implicitly is 1 segment
     in the Modflow 6 input."""
 
@@ -72,8 +66,8 @@ def test_get_options__no_segments(
         surface=elevation_fc,
         rate=rate_fc,
         depth=elevation_fc,
-        proportion_rate=proportion_rate_fc,
-        proportion_depth=proportion_depth_fc,
+        proportion_rate=None,
+        proportion_depth=None,
     )
 
     # Act
@@ -114,16 +108,60 @@ def test_get_options__with_segments(
     assert options["nseg"] == 4
 
 
-def test_get_bin_ds__no_segments(
+def test_init__error_no_segment_dim(
     rate_fc, elevation_fc, proportion_rate_fc, proportion_depth_fc
 ):
+    """Test without segment dimension in proportion arrays, should raise ValidationError.
+    Proportion array fixtures miss a segmnent dimension"""
+
+    with pytest.raises(ValidationError, match="segment"):
+        imod.mf6.Evapotranspiration(
+            surface=elevation_fc,
+            rate=rate_fc,
+            depth=elevation_fc,
+            proportion_rate=proportion_rate_fc,
+            proportion_depth=proportion_depth_fc,
+        )
+
+
+def test_init__error_only_one_proportion_var(
+    rate_fc, elevation_fc, proportion_rate_fc, proportion_depth_fc
+):
+    """Test with only one proportion variable, should raise ValueError."""
+
+    with pytest.raises(
+        ValueError,
+        match="Both 'proportion_rate' and 'proportion_depth' must both be provided",
+    ):
+        imod.mf6.Evapotranspiration(
+            surface=elevation_fc,
+            rate=rate_fc,
+            depth=elevation_fc,
+            proportion_rate=proportion_rate_fc,
+            proportion_depth=None,
+        )
+
+    with pytest.raises(
+        ValueError,
+        match="Both 'proportion_rate' and 'proportion_depth' must both be provided",
+    ):
+        imod.mf6.Evapotranspiration(
+            surface=elevation_fc,
+            rate=rate_fc,
+            depth=elevation_fc,
+            proportion_rate=None,
+            proportion_depth=proportion_depth_fc,
+        )
+
+
+def test_get_bin_ds__no_proportion_vars(rate_fc, elevation_fc):
     # Arrange
     evt = imod.mf6.Evapotranspiration(
         surface=elevation_fc,
         rate=rate_fc,
         depth=elevation_fc,
-        proportion_rate=proportion_rate_fc,
-        proportion_depth=proportion_depth_fc,
+        proportion_rate=None,
+        proportion_depth=None,
     )
 
     # Act
@@ -135,15 +173,13 @@ def test_get_bin_ds__no_segments(
         "surface",
         "rate",
         "depth",
-        "proportion_depth",
-        "proportion_rate",
     ]
 
     assert bin_ds.sizes == expected_sizes
     assert list(bin_ds.keys()) == expected_variables
 
 
-def test_get_bin_ds__with_segments(
+def test_get_bin_ds__with_proportion_vars(
     rate_fc, elevation_fc, proportion_rate_fc, proportion_depth_fc
 ):
     # Arrange